+++ datanode/networkhistory/service_test.go
@@ -379,12 +379,12 @@ func TestMain(t *testing.M) {
 		log.Infof("%s", goldenSourceHistorySegment[4000].HistorySegmentID)
 		log.Infof("%s", goldenSourceHistorySegment[5000].HistorySegmentID)
 
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[1000].HistorySegmentID, "QmdHV8ExVqvig51hDTtRyGuTKDTZQeu5QfCm5R98y3b8mp", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2000].HistorySegmentID, "QmXTJ8fm4CMR6hqzwFNifGyZHkTV1E81wC4UUgUTPspSt5", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2500].HistorySegmentID, "QmamvTjtH9tUM9ZYhArvY5nAra3J5vquGaAfhZqHK1uRn3", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[3000].HistorySegmentID, "Qma3kuUejJ5hGsdbSBctAUChiuA6kAwGQPwPhhUUSH23bf", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[4000].HistorySegmentID, "QmXNn2cvrHet87AWgeAviQt7L4Wm3TwdJ4du5kqArrHSrc", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[5000].HistorySegmentID, "QmevQk4GA31HN5HZ2byggZsnLizyyLuZyZeh82yTpvqPvH", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[1000].HistorySegmentID, "QmVWnyMKCqUgbEKgqu8A2TrbF4ZHcpZwA7md3drJnytM7w", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2000].HistorySegmentID, "QmZ1ermXvdg5Q46AUsjFbcRyua2CRyCAvzSnUDoTJS1ocy", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2500].HistorySegmentID, "QmXRbe6e3aSAoaHbmq6r2rT652JVi6iVatraRCx5YC3rHK", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[3000].HistorySegmentID, "QmUGEnbnaG6teZ34ZS7pLiz6kiEFioL8TCLu65nPVnWUis", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[4000].HistorySegmentID, "QmNbpm2UxCpPhoq5k9PdWhfWRPjFBu2dc2cDvTtJAG3fwY", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[5000].HistorySegmentID, "QmSRoiei5VNMPtzjGL4QwT8uShFCAwFDL8BaaFjc6Kcgqc", snapshots)
 	}, postgresRuntimePath, sqlFs)
 
 	if exitCode != 0 {